+++ ts/src/adapters/httpAdapter.ts
@@ -16,7 +16,7 @@
 import http from "http";
 import { HttpContext } from "../http-signature/HttpContext";
 
-import { Runtime } from "../sigruntime/runtime";
+import { Runtime } from "../runtime/runtime";
 import {
   Context, Request, Response, StringList
 } from "../http-signature/generated/generated";
