+++ ts/src/adapters/nextAdapter.test.ts
@@ -42,7 +42,7 @@ import { Context as PgContext, Request as PgRequest, Response as PgResponse, Str
 
 import { ScaleFunc, V1Alpha, Go } from "@loopholelabs/scalefile";
 import { HttpContext, HttpContextFactory } from "../http-signature/HttpContext";
-import { Runtime as SigRuntime, WasiContext } from "../sigruntime/runtime";
+import { Runtime as SigRuntime, WasiContext } from "../runtime/runtime";
 
 
 function getNewWasi(): WasiContext {
