+++ ts/src/runtime/cachedWasmInstance.ts
