+++ ts/src/runtime/sfunction.ts
