--- ts/src/sigtest.sh
@@ -1,1 +1,0 @@
-npm test -- sigruntime/runtime.test.ts
